Chat_with_docs/requirements.txt
anthropic
python-dotenv
python-docx
openai
anthropic
google-genai